Fix broken optimization in WordVec::to_raw
WordVec::to_raw returns the raw internal buffer directly, which may not be
correct in all cases, due to optimizations elsewhere.

The two cases in particular are as follows:

* write(N, 0) does not actually store N '0' bits in the buffer; instead, it
simply increments the internal length indicator.  As such, if the last write
to the stream wrote a zero value, to_raw() may be shorter than it ought to be.

* Other write operations increase the size by multiples of 32 bits.  Therefore,
the return value could be up to 3 bytes longer than it should have been.